EXP-13/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
pytest==8.3.2
pytest-cov==5.0.0